Switch to double-precision floats
There's no reason to use single-precision floats

+++ src/hslib/Benchmarking.hs
@@ -6,21 +6,21 @@ import Control.Concurrent (forkIO)
 
 -- mini benchmarking library
 
-elapsedMicros :: TimeSpec -> IO Float
+elapsedMicros :: TimeSpec -> IO Double
 elapsedMicros start = do
   end <- getTime Monotonic
   let elapsedNanos = toNanoSecs $ diffTimeSpec start end
-      elapsed = (fromIntegral elapsedNanos)/1e3 :: Float in
+      elapsed = (fromIntegral elapsedNanos)/1e3 :: Double in
     return elapsed
 
-timed :: IO a -> IO (a, Float)
+timed :: IO a -> IO (a, Double)
 timed act = do
   start <- getTime Monotonic
   r <- act
   totalTime <- elapsedMicros start
   return (r, totalTime)
 
-timeIt :: IO a -> IO Float
+timeIt :: IO a -> IO Double
 timeIt act = snd <$> timed act
 
 runInThread :: IO a -> IO (MVar a)
@@ -33,7 +33,7 @@ runInThread act = do
     putMVar m v
   return m
 
-timeAsync :: IO a -> IO (MVar Float)
+timeAsync :: IO a -> IO (MVar Double)
 timeAsync act = do
   start <- getTime Monotonic
   m <- newEmptyMVar
